requirements.txt
# Core dependencies
openai>=1.3.0
anthropic>=0.7.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0

# Vector stores and retrieval
faiss-cpu>=1.7.4
chromadb>=0.4.15
sentence-transformers>=2.2.2
rank-bm25>=0.2.2

# Reranking
sentence-transformers[cross-encoder]>=2.2.2

# API server
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Utilities
python-dotenv>=1.0.0
tiktoken>=0.5.2
numpy>=1.24.0
pandas>=2.1.0

# Database (for memory)
sqlalchemy>=2.0.23
# sqlite3 is built-in to Python, no need to install

# Evaluation
pytest>=7.4.0
pytest-asyncio>=0.21.0

